removed test break + altered test steps
Only proxies which work or were in the working list are updated. This means that unreliable proxies will never be updated and will go down the list with lastTested. Non-working proxy tester will first query result set to be ordered with most recent lastTested first. This will allow proxies which went online but then went offline for a short amount of time to be detected.