fixed tests + testing for race + fixed rand integration
Turns out in openproxy.space.go and shiftytr.github.go the isTest variable was being checked too early and no lists were being indexed. Now it is such that 1 list shall be indexed, to allow for testing all of the scraping code but to minimize the time taken by the test.